requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
fastapi==0.133.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0